pyproject.toml
[project]
name = "opening-hours"
version = "0.1.0"
dependencies = ["pytest", "pytest-approvals"]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
"" = "src"